+++ src/modules/mo-core/interfaces/redis-pub-sub-message.ts
@@ -71,3 +71,12 @@ export interface IRedisPubSubMessageAuth extends IRedisPubSubMessage {
   socketEvent: UserSocketEventEnum.AUTH_SERVICE;
   data: RedisPubSubMessageDataTypes<UserSocketDataEventEnum>[];
 }
+
+/**
+ * Hint: data type is UserSocketDataEventEnum
+ */
+export interface IRedisPubSubMessageNotification extends IRedisPubSubMessage {
+  redisChannel: 'mo_user';
+  socketEvent: UserSocketEventEnum.NOTIFICATION_SERVICE;
+  data: RedisPubSubMessageDataTypes<UserSocketDataEventEnum>[];
+}