+++ src/modules/mo-subscription-pool/data-transfer-objects/subscription.dto.ts
@@ -1,6 +1,6 @@
 import { Expose, Type } from 'class-transformer';
 import { BaseEntityDto } from '../../mo-core';
-import { CurrencyDto, PlanDto } from '../../mo-plan';
+import { CurrencyDto, PlanDto, PlanPublicDto } from '../../mo-plan';
 import { UserPublicDto } from '../../mo-user';
 import { SubscriptionStatusEnum } from '../enums';
 import { SubscriptionPoolDto } from './subscription-pool.dto';
@@ -19,7 +19,7 @@ export class SubscriptionDto extends BaseEntityDto {
   currency: CurrencyDto;
 
   @Expose()
-  plan: PlanDto;
+  plan: PlanPublicDto;
 
   @Expose()
   status: SubscriptionStatusEnum;